requirements.txt
vosk
sounddevice
paho-mqtt
pyserial
